requirements.txt
certifi==2018.11.29
chardet==3.0.4
decorator==4.0.11
Django==2.1.4
django-widget-tweaks==1.4.3
gunicorn==19.9.0
idna==2.8
imageio==2.4.1
moviepy==0.2.3.5
numpy==1.15.4
oauthlib==2.1.0
Pillow==5.4.0
python-twitch-client==0.6.0
pytz==2018.7
requests==2.11.1
requests-oauthlib==1.0.0
requests-toolbelt==0.7.0
six==1.12.0
tqdm==4.11.2
twitter==1.18.0
urllib3==1.24.1
